requirements.txt
pikepdf
pymupdf
Pillow
numpy
reportlab
